requirements.txt
# 兼容旧的 pip 安装方式；主依赖定义已迁移到 pyproject.toml / uv.lock。
moviepy==2.1.2
streamlit==1.45.0
edge_tts==7.2.7
fastapi==0.115.6
uvicorn==0.32.1
openai==1.56.1
faster-whisper==1.1.0
loguru==0.7.3
google.generativeai==0.8.6
dashscope==1.20.14
g4f==0.5.2.2
azure-cognitiveservices-speech==1.41.1
redis==5.2.0
python-multipart==0.0.19
pyyaml==6.0.3
requests==2.33.1
socksio==1.0.0
pydub==0.25.1

# ── SaaS Layer Dependencies ──
sqlalchemy[asyncio]==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
celery[redis]==5.4.0
stripe==11.4.1
PyJWT==2.10.1
python-dotenv==1.0.1
boto3==1.35.86
passlib[bcrypt]==1.7.4

